Add an example binary for testing against TUN interface
This allows us to evaluate at least some of the PseudoTCP stack without
having to run it inside of an Android App/Emulator.

It is *not* identical to running it as an Android App, in particular the
socket protection pieces and DNS are subtly different so it's important
to recognize that this does not comprise a full parity test environment.