Refactor into PseudoTCP struct
This essentially takes all of the variables that were previously
"package level" and moves them inside of a PseudoTCP struct.
Correspondingly, it takes all of the "package level" functions and makes
them methods of the PseudoTCP struct. This facilitates simpler
creation, configuration, initialization as well as allows multiple
instances of the PseudoTCP struct to be instantiated in parallel if
necessary.